feat(ui): rename share component
