+++ src/ui/components/ShareConnection/ShareConnection.scss
