+++ src/ui/components/ShareConnection/index.ts
@@ -1,0 +1,1 @@
+export * from "./ShareConnection";
