--- src/ui/components/ShareIdentifier/index.ts
@@ -1,1 +1,0 @@
-export * from "./ShareIdentifier";
